add dyn cfg replace_out_of_range_with_max_range
Add a new dynamic configure option to replace out-of-range data with
whatever the maximum range is set to. This is useful for implementing
clearing-on-invalid in a costmap context by setting the max range
beyond the max range used as the maximum obstacle range in a costmap.
The default is disabled.