force invalid intensity with replaced data
When replacing data with max range (such as when the new dynamic
configure option to replace out-of-range data is enabled), set
the intensity to `nanf("")`. This more correctly identifies the
data as invalid data that was retained for using in ray-tracing
through invalid returns. Also, by making the intensity NAN, the
cloud can be easily filtered into valid/invalid data using a
PCL PassThrough filter.