input: synchronize hardware timestamps if no GPS
If there is no GPS clock available, synchronize the hardware time
stamp to the system clock using an exponential moving average of the
difference between the hardware clock and system clock. Note that
this does not take into account the average latency. To offset
average latency, it should be measured and the existing time_offset
parameter set.