correctly handle azimuth roll over
Simlify and correct the code that handles azimuth roll over.
The azimuth angle is transmitted as an unsigned 16-bit integer
of 100ths of degrees (0 to 35999). The previous code incorrectly
assumed that the modulus does not return negative, which it does.
Correct and simplify the code by no longer using modulus, as it
is unncessary and expensive compared to simply checking for a
wrap and adding the 36000 correction factor. This is simple as
it is safe to assume the azimuth only decreases when wrapping.